- Model Friendship przeniesiony do folderu auth - merge z masterem - drobne zmiany w sciezkach w docker-compose.yml bo inaczej sie wszystko wykrzaczało
