+++ cityfeel/auth/models.py
@@ -1,5 +1,7 @@
 from django.contrib.auth.models import AbstractUser
 from django.db import models
+from django.conf import settings
+from django.db.models import Q, F
 
 
 def user_avatar_upload_path(instance, filename):
@@ -41,3 +43,36 @@ class CFUser(AbstractUser):
         if self.avatar:
             return self.avatar.url
         return None
+
+class Friendship(models.Model):
+    STATUS_CHOICES = [
+        ('pending', 'Pending'),
+        ('accepted', 'Accepted'),
+        ('rejected', 'Rejected'),
+    ]
+
+    user = models.ForeignKey(
+        settings.AUTH_USER_MODEL,
+        on_delete=models.CASCADE,
+        related_name='friendships_initiated'
+    )
+    friend = models.ForeignKey(
+        settings.AUTH_USER_MODEL,
+        on_delete=models.CASCADE,
+        related_name='friendships_received'
+    )
+    status = models.CharField(
+        max_length=10,
+        choices=STATUS_CHOICES,
+        default='pending'
+    )
+    created_at = models.DateTimeField(auto_now_add=True)
+
+    class Meta:
+        constraints = [
+            models.UniqueConstraint(fields=['user', 'friend'], name='unique_friendship'),
+            models.CheckConstraint(check=~Q(user=F('friend')), name='users_cannot_friend_themselves')
+        ]
+
+    def __str__(self):
+        return f"{self.user} -> {self.friend} ({self.status})"
