+++ docker-compose.yml
@@ -27,7 +27,7 @@ services:
       dockerfile: ./docker/web/Dockerfile
     container_name: cityfeel_web
     restart: unless-stopped
-    command: uv run python manage.py runserver 0.0.0.0:8000
+    command: /app/.venv/bin/python manage.py runserver 0.0.0.0:8000 #z poprzednią ścieżką mi nie działało, nie wiem ocb. Ta zmiana załatawiła problem
     ports:
       - "8000:8000"
     env_file:
@@ -40,6 +40,7 @@ services:
         condition: service_healthy
     environment:
       - DB_HOST=postgres
+      - PYTHONPATH=/app/cityfeel #bez tego tez nie działało - problem ze ścieżką.
     profiles: ["server"]
 
 volumes:
