--- social/admin.py
@@ -1,8 +1,0 @@
-from django.contrib import admin
-from .models import Friendship
-
-@admin.register(Friendship)
-class FriendshipAdmin(admin.ModelAdmin):
-    list_display = ('user', 'friend', 'status', 'created_at')
-    list_filter = ('status',)
-    search_fields = ('user__username', 'friend__username')
